refactor: Add SectionListTile component for displaying sections in home_view.dart
This commit adds a new component called SectionListTile in the section_list_tile.dart file. The SectionListTile component is used to display a section title, an icon, and a list of items. It improves the modularity and reusability of the code by encapsulating the logic for rendering a section in a separate component.